LayoutTests/http/tests/security/contentSecurityPolicy/audio-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load media from 'http://localhost:8000/resources/balls-of-the-orient.aif' because it violates the following Content Security Policy directive: "media-src http://127.0.0.1:8000/resources/redirect.php".

ALERT: PASS
